refactor: Project -> ProjectCard로 컴포넌트 이름 변경
