+++ src/pages/Projects/components/style/projectCard.ts
